+++ aws/last_commit_hash
@@ -1,1 +1,1 @@
-590ab6fb9c89aa05df6d4a31cb379cfd6213ea2e
+c1ba93553bfaa9816e21ce21a0421aa282307f1c
