--- web/styles/navbar_light.css
@@ -1,29 +1,0 @@
-#navbar {
-    margin-right: -10px;
-    padding-right: 10px;
-    padding-left: 10px;
-
-}
-
-#navbar:hover #navbar_links {
-    background-color: rgba(255, 255, 255, 0.5);
-    backdrop-filter: blur(10px);
-    -webkit-backdrop-filter: none;
-    border-radius: 20px;
-    height: auto;
-    width: auto;
-    opacity: 1;
-    transition: all 0.3s ease-in;
-    -webkit-transition: all 0.3s ease-in;
-}
-
-
-#navbar_text {
-    color: var(--moonstone);
-    text-shadow: 0 0 10px rgb(255, 255, 255);
-}
-
-#navbar_links a {
-    text-shadow: 0px 0px 10px rgb(255, 255, 255);
-    color: rgb(73, 73, 73)
-}
